Create textNaiveBayes.py
Naive Bayes Classifier for text processing. Two examples are included.

In the first example, Probabilities of five words sent by two individuals in e-mails are known and populated in a pandas data frame. 
Based on three test words (present in the vocabulary), a naive bayes classifier identifies the sender of the mail

In the second example, product sentiments are either assessed positive or negative. Probabilities of seven words are known.
Test data has words which are not in vocabulary are assigned equal probabilities.
As a first step, test data words present in the original dataset are subset and stored as a new data frame (newDF)
Another data frame is created for words present in test data but not present in original dataset is called as out of vocabulary data frame (oovDF)
newDF and oovDF are concatenated into a data frame called as merged
Calculations are done to check the review is positive or negative