Create zipfMandel.py
Zipf's law states that given a large sample of words used, the frequency of any word is inversely proportional to its rank in the frequency table. . Thus the most frequent word will occur about twice as often as the second most frequent word, three times as often as the third most frequent word, etc.  There is a constant k such that frequency multiplied by rank is a constant. When plotted, Zipf's law can display the frequency distribution of words in a corpus

Mandelbrot's approximation deals with the estimation of size of vocabulary of the corpus given the total number of tokens.

emma austen corpus available in nltk is considered. The top 10 frequent words in the corpus are considered and plotted

Significant portion of the code is provided by Prof Ramaseshan in his NPTEL anlp course. Additional codes are included to plot zipf's law, mandelbrot's approximation.

Code can be altered especially in the number of frequent words to be considered. Ten words are considered in this example. 
It can be observed that Zipf's law provides better approximation for frequency of the words at the top end of the curve when compared to mandelbrot's approximation
with beta = 2.7 and alpha = 1